requirements.txt
discord.py==2.0.0
aiohttp==3.7.4
python-dotenv==0.19.2
